requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
filelock==3.7.1
huggingface-hub==0.8.1
idna==3.3
joblib==1.1.0
nltk==3.7
numpy
packaging==21.3
pandas
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
scipy
six==1.16.0
tokenizers==0.12.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=1.10
tornado==6.2
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
